server/requirements.txt
fastapi[standard]
yt-dlp
uvicorn
python-multipart
ffmpeg-python
google-api-python-client
python-dotenv
